Fix creating secret by forms
In the previous change we expected that when one sends message from a
simple html form, the header will be application/x-www-form-urlencoded,
which was right and in that case we were searching for the secret under
the `secret` key of the form post data.

However it breaks the interface when it's used with curl and the simple
curl usage was a main point of the tool. (`curl -d secret
https://sekret.link/api`)

With the previous change the behavior has been broken because curl uses
application/x-www-form-urlencoded format when the `-d` flag used, but it
does not set any name for the data. So the secret were not extracted
from the post body.

So if client want to set a name for the data - and HTML forms won't send
the field value if it has no name - the content type must be
multipart/form-data.
That way even images or anything can be sent as a secret.