Add delete route
Deleting entry requires a delete key, which provided to the user only
when the entry created.
The UUID of the entry and the delete key must be matching to delete the
entry.

It was considered to encrypt a predefined secret with the same key what
we use to encrypt the actual data, but that would weaken the encryption,
by if the the database and the delete key secret would be stolen
somehow, it would make it possible to figure out the key with bruteforce
attack - however still wouldn't be easy.

Using the delete key, the encrypted data is still safe.