Do not clear the storage in tests
Some were failing, because the newly created item were not there in the
next step, when GetMeta were called. It happened if the tests were
called with `go test ./...` command to test all of the packages.
It turned out that if we test the packages like this, go test runs the
tests paralelly, even if the `-parallel` option set to 1.

Since the database cleanup was just optional, remove the Clean call.