delete secret after the read operation confirmed
When we read the secret from postgresql storage because we had the good
ID we also deleted it from the database.
But if the decryption key was incorrect the user were not be able to
read the secret when she retired to read it with the good key.
It casues confusion, as I would expect that we delete the secret if the
read was successfull.

So introduced a confirmed reader interface. Implementer should wait for
the confirm channel to confirm that the read happened, and mark it as
read only after that.