Comprehensive Code Restructuring for Improved Readability and Maintainability (#1)
introduces a comprehensive restructuring of our codebase with the primary aim
of enhancing readability and maintainability. Here's a detailed breakdown of
the changes:

- Internal Package: We have relocated all existing code to an 'internal'
package. This strategic move leaves the 'main' package solely for the main
function, thereby streamlining its purpose and making it easier to locate key
components of our application.
- Durable Layer: A new 'durable' layer has been introduced. This layer is
specifically designed to manage database connections.
- Model Layer: We have added a new 'model' layer. This layer is responsible for
all data-related operations, providing a clear and distinct location for data
handling and manipulation.
- Service Layer: A 'service' layer has been implemented to handle all business
logic. This separation of business rules from other aspects of the code
improves readability and makes it easier to modify or extend business rules in
the future.
- Parser and View Layers: To further enhance the separation of concerns, we
have introduced a 'parser' layer for request parsing and a 'view' layer for
response formatting. This ensures that each aspect of request handling and
response generation is handled by a dedicated component, improving
maintainability.
- Routing: We have updated our routing to be handled by Go 1.22's HTTP router.
This provides us with more robust and efficient routing capabilities.
- Additional Tests: We have added additional tests using Hurl to test the API,
ensuring our application's robustness and reliability.
- Testify for Testing: We have adopted Testify for testing, providing a more
flexible and powerful testing framework.

These changes represent a significant step forward in the structure of our
codebase, and we believe they will greatly improve both the readability and
maintainability of our code.