feat(tests): refactor tests to use testify assertions
This commit refactors the tests in api_test.go to use the testify
assertion library for cleaner, more readable tests. It also introduces
a new function, FromString, in key.go to handle keys represented as
strings. This function checks the length of the string and calls
FromHex if the length is 64, otherwise it returns an ErrorInvalidKey.